extract getTableOperationHandler to the table_helper utils
