--- js/ui/html_editor/modules/tableOperations.js
@@ -1,11 +1,0 @@
-export const getTableOperationHandler = (quill, operationName, ...rest) => {
-    return () => {
-        const table = quill.getModule('table');
-
-        if(!table) {
-            return;
-        }
-        quill.focus();
-        return table[operationName](...rest);
-    };
-};
